requirements.txt
Flask==3.0.0
Pillow==10.1.0
Flask-Limiter==3.5.0
python-magic-bin==0.4.14
